+++ NinjaGaiden/GameComponents/PinkWitch.h
@@ -18,7 +18,7 @@ class PinkWitch :public Enemy
 	static vector<Animation *> animations;
 public:
 	PinkWitch();
-	PinkWitch(int posx, int posy);
+	PinkWitch(float posx, float posy);
 	void LoadResources();
 	void Idle();
 	void Walk();
